+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Changes for lovely-kafka-backup
 
+## Unreleased
+
+### Feature
+
+- Breaking change: `readAll()` in `RecordStreamReader` returns a cold `Flow` instead of a `List`.
+
 ## 2023-09-13 / 0.1.0
 
 - Initial release